+++ go.mod
@@ -3,16 +3,19 @@ module github.com/gotzmann/llama.go
 go 1.20
 
 require (
-	github.com/gotzmann/llama.go/llama v0.0.0-20230330223640-1e49860bd552
 	github.com/gotzmann/llama.go/ml v0.0.0-20230330223640-1e49860bd552
 	github.com/jessevdk/go-flags v1.5.0
 	github.com/mattn/go-colorable v0.1.13
 	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db
+	github.com/schollz/progressbar/v3 v3.13.1
+	github.com/x448/float16 v0.8.4
+	golang.org/x/exp v0.0.0-20230321023759-10a507213a29
 )
 
 require (
-	github.com/mattn/go-isatty v0.0.16 // indirect
-	github.com/x448/float16 v0.8.4 // indirect
-	golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
+	github.com/mattn/go-runewidth v0.0.14 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
 	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/term v0.6.0 // indirect
 )
