--- go.work
@@ -1,7 +1,0 @@
-go 1.20
-
-use (
-	.
-	./llama
-	./ml
-)
