--- llama/go.mod
@@ -1,15 +1,0 @@
-module github.com/gotzmann/llama.go/llama
-
-go 1.20
-
-require (
-	github.com/gotzmann/llama.go/ml v0.0.0-20230409185708-181ed7a95d40 // indirect
-	github.com/mattn/go-runewidth v0.0.14 // indirect
-	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
-	github.com/rivo/uniseg v0.2.0 // indirect
-	github.com/schollz/progressbar/v3 v3.13.1 // indirect
-	github.com/x448/float16 v0.8.4 // indirect
-	golang.org/x/exp v0.0.0-20230321023759-10a507213a29 // indirect
-	golang.org/x/sys v0.6.0 // indirect
-	golang.org/x/term v0.6.0 // indirect
-)
