--- ml/go.mod
@@ -1,5 +1,0 @@
-module github.com/gotzmann/llama.go/ml
-
-go 1.20
-
-require github.com/x448/float16 v0.8.4
